Switch to bf16 training: FP8 backward gradient overflow in deep models
FP8 e4m3fn backward pass produces NaN due to gradient magnitude overflow:
- Gradient values grow exponentially through matmul chain in backprop
- Even He-initialized weights overflow at ~6 matmul layers
- Our model has ~142 matmuls (31 CNN blocks + 16 mixer layers)
- FP8 range [-448,448] with ~5 orders of magnitude dynamic range
  cannot accommodate gradient magnitudes varying 10+ orders across layers
- Loss scaling cannot fix this: no single scale works for all layers
- proper_element_type=f32 works for dot_general but not conv backward
- ANY_F8_ANY_F8_F32 gives 'Unsupported dot precision algorithm' for Linear

Proper FP8 training requires per-tensor dynamic scaling (NVIDIA Transformer
Engine approach). Keep bf16 for now; fp8 model safety code preserved for
future per-tensor scaling implementation.

+++ train.py
@@ -34,10 +34,13 @@ from tensorboardX import SummaryWriter
 from absl import app
 
 MODEL_DTYPE = jnp.float32
-FORWARD_DTYPE = jnp.float8_e4m3fn
-BACKWARD_DTYPE = jnp.float8_e4m3fn
-# FORWARD_DTYPE = jnp.bfloat16
-# BACKWARD_DTYPE = jnp.bfloat16
+# FP8 training is not viable without per-tensor dynamic scaling (NVIDIA TE approach).
+# FP8 backward gradient overflow starts at ~6 matmul layers; our model has ~142.
+# See session plan for full analysis.  Keep bf16 for training stability.
+# FORWARD_DTYPE = jnp.float8_e4m3fn
+# BACKWARD_DTYPE = jnp.float8_e4m3fn
+FORWARD_DTYPE = jnp.bfloat16
+BACKWARD_DTYPE = jnp.bfloat16
 # FORWARD_DTYPE = jnp.float32
 # BACKWARD_DTYPE = jnp.float32
